DRILL-6155: Refactoring the existing code.
