+++ exec/java-exec/src/main/java/org/apache/drill/exec/physical/config/AbstractMuxExchange.java
@@ -29,6 +29,7 @@ import org.apache.drill.exec.physical.EndpointAffinity;
 import org.apache.drill.exec.physical.MinorFragmentEndpoint;
 import org.apache.drill.exec.physical.base.AbstractExchange;
 import org.apache.drill.exec.physical.base.PhysicalOperator;
+import org.apache.drill.exec.physical.base.Receiver;
 import org.apache.drill.exec.physical.base.Sender;
 import org.apache.drill.exec.planner.fragment.ParallelizationInfo;
 import org.apache.drill.exec.proto.CoordinationProtos.DrillbitEndpoint;
@@ -51,6 +52,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * for the 10 fragments is 300 instead of earlier number 10*300.
  */
 public abstract class AbstractMuxExchange extends AbstractExchange {
+  private static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(AbstractMuxExchange.class);
 
   // Ephemeral info used when creating execution fragments.
   protected Map<Integer, MinorFragmentEndpoint> senderToReceiverMapping;
@@ -88,6 +90,19 @@ public abstract class AbstractMuxExchange extends AbstractExchange {
     }
 
     return new SingleSender(receiverMajorFragmentId, receiver.getId(), child, receiver.getEndpoint());
+  }
+
+  protected final List<MinorFragmentEndpoint> getSenders(int minorFragmentId) {
+    createSenderReceiverMapping();
+
+    List<MinorFragmentEndpoint> senders = receiverToSenderMapping.get(minorFragmentId);
+
+    logger.debug("Minor fragment {}, receives data from following senders {}", minorFragmentId, senders);
+    if (senders == null || senders.size() <= 0) {
+      throw new IllegalStateException(String.format("Failed to find senders for receiver [%d]", minorFragmentId));
+    }
+
+    return senders;
   }
 
   protected void createSenderReceiverMapping() {
