+++ exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/HashPrelUtil.java
@@ -19,19 +19,15 @@ package org.apache.drill.exec.planner.physical;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.calcite.rel.type.RelDataType;
+import org.apache.calcite.rex.RexBuilder;
 import org.apache.calcite.rex.RexNode;
 import org.apache.drill.common.expression.ExpressionPosition;
 import org.apache.drill.common.expression.FieldReference;
 import org.apache.drill.common.expression.FunctionCall;
 import org.apache.drill.common.expression.LogicalExpression;
 import org.apache.drill.common.expression.ValueExpressions;
-import org.apache.drill.common.logical.data.NamedExpression;
-import org.apache.drill.common.types.TypeProtos;
-import org.apache.drill.common.types.Types;
-import org.apache.drill.exec.expr.DirectExpression;
-import org.apache.drill.exec.expr.ValueVectorReadExpression;
 import org.apache.drill.exec.planner.physical.DrillDistributionTrait.DistributionField;
-import org.apache.drill.exec.record.TypedFieldId;
+import org.apache.drill.exec.planner.sql.DrillSqlOperator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -62,6 +58,21 @@ public class HashPrelUtil {
           return new FunctionCall(funcName, inputFiled, ExpressionPosition.UNKNOWN);
         }
       };
+
+  public static class RexNodeBasedHashExpressionCreatorHelper implements HashExpressionCreatorHelper<RexNode> {
+    private final RexBuilder rexBuilder;
+
+    public RexNodeBasedHashExpressionCreatorHelper(RexBuilder rexBuilder) {
+      this.rexBuilder = rexBuilder;
+    }
+
+    @Override
+    public RexNode createCall(String funcName, List<RexNode> inputFields) {
+      final DrillSqlOperator op =
+              new DrillSqlOperator(funcName, inputFields.size(), true, false);
+      return rexBuilder.makeCall(op, inputFields);
+    }
+  }
 
   // The hash32 functions actually use hash64 underneath.  The reason we want to call hash32 is that
   // the hash based operators make use of 4 bytes of hash value, not 8 bytes (for reduced memory use).
