+++ exec/java-exec/src/main/java/org/apache/drill/exec/planner/physical/HashToRandomExchangePrel.java
@@ -18,10 +18,17 @@
 package org.apache.drill.exec.planner.physical;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.List;
 
+import com.google.common.collect.Lists;
 import org.apache.calcite.linq4j.Ord;
 
+import org.apache.calcite.rel.type.RelDataType;
+import org.apache.calcite.rel.type.RelDataTypeField;
+import org.apache.calcite.rex.RexBuilder;
+import org.apache.calcite.rex.RexNode;
+import org.apache.calcite.rex.RexUtil;
 import org.apache.drill.exec.physical.base.PhysicalOperator;
 import org.apache.drill.exec.physical.config.HashToRandomExchange;
 import org.apache.drill.exec.planner.cost.DrillCostBase;
@@ -35,6 +42,7 @@ import org.apache.calcite.plan.RelOptCluster;
 import org.apache.calcite.plan.RelOptCost;
 import org.apache.calcite.plan.RelOptPlanner;
 import org.apache.calcite.plan.RelTraitSet;
+import org.apache.drill.exec.server.options.OptionManager;
 
 
 public class HashToRandomExchangePrel extends ExchangePrel {
@@ -110,6 +118,75 @@ public class HashToRandomExchangePrel extends ExchangePrel {
         pw.item("dist" + ord.i, ord.e);
       }
     return pw;
+  }
+
+  /**
+   * This method creates a new UnorderedMux and Demux exchanges if mux operators are enabled.
+   * @param child input to the new Unordered[Mux/Demux]Prel or new HashToRandomExchange node.
+   * @param options options manager to check if mux is enabled.
+   */
+  @Override
+  public Prel constructMuxPrel(Prel child, OptionManager options) {
+    boolean isMuxEnabled = options.getOption(PlannerSettings.MUX_EXCHANGE.getOptionName()).bool_val;
+    Prel newPrel = child;
+
+    final List<String> childFields = child.getRowType().getFieldNames();
+
+    List <RexNode> removeUpdatedExpr = null;
+
+    if (isMuxEnabled) {
+      // Insert Project Operator with new column that will be a hash for HashToRandomExchange fields
+      final List<DistributionField> distFields = getFields();
+      final List<String> outputFieldNames = Lists.newArrayList(childFields);
+      final RexBuilder rexBuilder = getCluster().getRexBuilder();
+      final List<RelDataTypeField> childRowTypeFields = child.getRowType().getFieldList();
+
+      final HashPrelUtil.HashExpressionCreatorHelper<RexNode> hashHelper =
+                                    new HashPrelUtil.RexNodeBasedHashExpressionCreatorHelper(rexBuilder);
+
+      final List<RexNode> distFieldRefs = Lists.newArrayListWithExpectedSize(distFields.size());
+      for (int i=0; i<distFields.size(); i++) {
+        final int fieldId = distFields.get(i).getFieldId();
+        distFieldRefs.add(rexBuilder.makeInputRef(childRowTypeFields.get(fieldId).getType(), fieldId));
+      }
+
+      final List <RexNode> updatedExpr = Lists.newArrayListWithExpectedSize(childRowTypeFields.size());
+      removeUpdatedExpr = Lists.newArrayListWithExpectedSize(childRowTypeFields.size());
+      for (RelDataTypeField field : childRowTypeFields) {
+        RexNode rex = rexBuilder.makeInputRef(field.getType(), field.getIndex());
+        updatedExpr.add(rex);
+        removeUpdatedExpr.add(rex);
+      }
+
+      outputFieldNames.add(HashPrelUtil.HASH_EXPR_NAME);
+      final RexNode distSeed = rexBuilder.makeBigintLiteral(BigDecimal.valueOf(HashPrelUtil.DIST_SEED)); // distribution seed
+      updatedExpr.add(HashPrelUtil.createHashBasedPartitionExpression(distFieldRefs, distSeed, hashHelper));
+
+      RelDataType rowType = RexUtil.createStructType(getCluster().getTypeFactory(), updatedExpr, outputFieldNames);
+
+      ProjectPrel addColumnprojectPrel = new ProjectPrel(child.getCluster(), child.getTraitSet(), child, updatedExpr, rowType);
+
+      newPrel = new UnorderedMuxExchangePrel(addColumnprojectPrel.getCluster(), addColumnprojectPrel.getTraitSet(),
+              addColumnprojectPrel);
+    }
+
+    newPrel = new HashToRandomExchangePrel(getCluster(), getTraitSet(), newPrel, getFields());
+
+    if (options.getOption(PlannerSettings.DEMUX_EXCHANGE.getOptionName()).bool_val) {
+      HashToRandomExchangePrel hashExchangePrel = (HashToRandomExchangePrel) newPrel;
+      // Insert a DeMuxExchange to narrow down the number of receivers
+      newPrel = new UnorderedDeMuxExchangePrel(getCluster(), getTraitSet(), hashExchangePrel, hashExchangePrel.getFields());
+    }
+
+    if (isMuxEnabled) {
+      // remove earlier inserted Project Operator - since it creates issues down the road in HashJoin
+      RelDataType removeRowType = RexUtil.createStructType(newPrel.getCluster().getTypeFactory(), removeUpdatedExpr, childFields);
+
+      ProjectPrel removeColumnProjectPrel = new ProjectPrel(newPrel.getCluster(), newPrel.getTraitSet(),
+                                                            newPrel, removeUpdatedExpr, removeRowType);
+      return removeColumnProjectPrel;
+    }
+    return newPrel;
   }
 
   @Override
